refactor: fix installer to not run on mac

+++ install.sh
@@ -1,14 +1,18 @@
 #!/usr/bin/env bash
 
-# Set up repositories
-curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
-echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
-curl -sL https://deb.nodesource.com/setup_14.x -o nodesource_setup.sh
-sudo bash nodesource_setup.sh
-rm -rf nodesource_setup.sh
+SYSTEM=$3
 
-# Install node
-sudo apt install nodejs -y
+if [ "$SYSTEM" != "mac" ]; then
+    # Set up repositories
+    curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | sudo apt-key add -
+    echo "deb https://dl.yarnpkg.com/debian/ stable main" | sudo tee /etc/apt/sources.list.d/yarn.list
+    curl -sL https://deb.nodesource.com/setup_14.x -o nodesource_setup.sh
+    sudo bash nodesource_setup.sh
+    rm -rf nodesource_setup.sh
 
-# Install yarn
-sudo apt install yarn -y
+    # Install node
+    sudo apt install nodejs -y
+
+    # Install yarn
+    sudo apt install yarn -y
+fi
